+++ modules/base/security.nix
@@ -6,10 +6,10 @@
 {
   ## Security & Hardening Settings ##
   config = {
-    # Protocols
-    programs.firejail.enable = true;
     security = {
       protectKernelImage = false; # To enable Hibernation
+
+      # Security Module
       apparmor = {
         enable = true;
         killUnconfinedConfinables = true;
@@ -17,21 +17,26 @@
       };
     };
 
+    # Block Junk Sites
+    networking.stevenblack.enable = true;
+
+    # Sandbox
+    programs.firejail.enable = true;
+
     boot = {
-      # Parameters
+      # Hardening
       kernelParams = [
         "page_alloc.shuffle=1"
         "page_poison=1"
         "slub_debug=FZP"
       ];
 
-      # Flags
       kernel.sysctl = {
         # Kernel
         "kernel.ftrace_enabled" = false;
         "kernel.kexec_load_disabled" = true;
         "kernel.kptr_restrict" = lib.mkOverride 500 2;
-        "kernel.sysrq" = 0;
+        "kernel.sysrq" = 176;
         "kernel.yama.ptrace_scope" = lib.mkOverride 500 1;
 
         # Network
@@ -93,22 +98,5 @@
         "ufs"
       ];
     };
-
-    ## Block Junk Sites
-    networking.extraHosts =
-      builtins.readFile (
-        pkgs.fetchurl {
-          # Shady Sites
-          url = "https://raw.githubusercontent.com/shreyasminocha/shady-hosts/a5647df22b0dc5ff6c866f21ee2d8b588682626a/hosts";
-          sha256 = "sha256-f6xCphzFHFYfOhMv+488amAMy+rSyHxQs/21nyvBtiU=";
-        }
-      )
-      + builtins.readFile (
-        pkgs.fetchurl {
-          # Crypto Scams
-          url = "https://raw.githubusercontent.com/MetaMask/eth-phishing-detect/78e727318a77fd62521d220e25871cdb65b8f00d/src/hosts.txt";
-          sha256 = "sha256-b3HvaLxnUJZOANUL/p+XPNvu9Aod9YLHYYtCZT5Lan0=";
-        }
-      );
   };
 }